# rust: remove copy in GCS reading